convert Image and TextureImage to use size_t

+++ include/GameLibrary/Exception/Graphics/ImageOutOfBoundsException.hpp
@@ -10,30 +10,30 @@ namespace fgl
 	{
 	public:
 		/*! The out of bounds index*/
-		unsigned int index;
+		size_t index;
 		/*! the x pixel*/
-		unsigned int x;
+		size_t x;
 		/*! the y pixel*/
-		unsigned int y;
+		size_t y;
 		
 		/*! the size of the image (width * height)*/
-		unsigned int size;
+		size_t size;
 		/*! the width of the image*/
-		unsigned int width;
+		size_t width;
 		/*! the height of the image*/
-		unsigned int height;
+		size_t height;
 		
 		/*! Constructs an exception for an x,y coordinate, and a width and height of an image.
 			\param x the x pixel
 			\param y the y pixel
 			\param width the width of the image
 			\param height the height of the image*/
-		ImageOutOfBoundsException(unsigned int x, unsigned int y, unsigned int width, unsigned int height);
+		ImageOutOfBoundsException(size_t x, size_t y, size_t width, size_t height);
 		/*! Constructs an exception for the pixel index and a width and height of an image.
 			\param index the pixel index
 			\param width the width of the image
 			\param height the height of the image*/
-		ImageOutOfBoundsException(unsigned int index, unsigned int width, unsigned int height);
+		ImageOutOfBoundsException(size_t index, size_t width, size_t height);
 		
 		/*! \copydoc fgl::Exception::getClassName()const*/
 		virtual String getClassName() const override;
